+++ src/address.h
@@ -5,14 +5,14 @@
 #define MAX_ADDRESS_SIZE 1000 /* the maximum size for network address */
 
 typedef struct {
-  long pid;
+  int pid;
   unsigned int port_in;
   unsigned int port_out;
   char *address;  
 } Address;
 
 Address *make_address_object(); 
-Address *create_address(Address *addr, char *address, int port_in, int port_out, long pid); 
+Address *create_address(Address *addr, char *address, int port_in, int port_out, int pid); 
 void free_address_object(Address *addr); 
 int serialize_address(char *buffer, Address *addr); 
 int deserialize_address(char *buffer, Address *addr);
